Final Submission
Includes new skybox, lighting, condition cheering animations, changed laser
