--- Assets/NewBehaviourScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9282cca2a2bb4b0469b956ae5ffc4494
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
