+++ Assets/Resources/OculusPlatformSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 663acccd2593dbc41a82bc331977307f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
