+++ Assets/Scripts/GameStartTriggerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6eb67e12927bbc143ab4567ec6526eca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
