+++ Assets/WildWestAssets/New Lighting Settings 1.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 024a81bb9c0112a4a90caad38c949230
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
