+++ Assets/WildWestAssets/jumer.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61f05eea3f67540459776d5e9c896911
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
