+++ Assets/WildWestAssets/sky.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f68f85b766454794a9a84f0eb695e2e1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
