+++ Assets/WildWestAssets/sky/sky.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e21f80c2e26beaa409b7205bcdf8edde
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
