+++ Assets/WildWestAssets/yt5s.io - Cold wind sound effect (320 kbps).mp3.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: dfced152f8b6b094d95b6ac968c0f03a
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
