+++ UserSettings/EditorUserSettings.asset
@@ -24,16 +24,16 @@ EditorUserSettings:
       value: 5352035655515e025f5e5e2348705d44411519282f2a75697b784b62b5b4646f
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 520755525d0d5c030f0c0f7540750e44144e4a737d7023332f2a4e36b0b46c61
-      flags: 0
-    RecentlyUsedSceneGuid-7:
       value: 57080100000d5159550b0e76447a5a44444f482f742b753374784c31e7b2616a
       flags: 0
-    RecentlyUsedSceneGuid-8:
+    RecentlyUsedSceneGuid-7:
       value: 5654560250075a035c5f5e2343705944404e4c737c2d77667978496be3e63260
       flags: 0
-    RecentlyUsedSceneGuid-9:
+    RecentlyUsedSceneGuid-8:
       value: 5655050200030d58595e5826487a074444164f7a2a7024637e2a4860b3e3656f
+      flags: 0
+    RecentlyUsedSceneGuid-9:
+      value: 520755525d0d5c030f0c0f7540750e44144e4a737d7023332f2a4e36b0b46c61
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
